requirements.txt
Django==4.1.5
djangorestframework==3.14.0
daphne==4.0.0
channels==4.0.0
twilio==7.16.1
django-phonenumber-field==7.0.2
phonenumbers==8.13.4
black==22.12.0
requests==2.28.2
python-dotenv==0.21.0
django-shortuuidfield==0.1.3
psycopg2-binary==2.9.5
django-cors-headers==3.13.0
channels-redis==4.0.0